+++ .eslintrc.js
@@ -2,9 +2,7 @@
  * eslint配置
  */
 module.exports = {
-  extends: require.resolve('@jianghe/sand-lint/react'),
-  rules: {},
-  env: {},
+  extends: require.resolve('@jianghe/sand-lint/react/prettier'),
   settings: {
     // 忽略别名导致的找不到模块报错
     'import/resolver': {
@@ -13,4 +11,5 @@ module.exports = {
       },
     },
   },
+  rules: {},
 };