+++ packages/sand-build/bin/sand-build.js
@@ -15,7 +15,7 @@ program
   // build的子命令
   .option('-w, --watch', '开启监听')
   .option('-t, --type <type>', '构建类型（lib||mob||pc||demo）')
-  .option('-e, --env <env>', '构建方式（prod||dev）')
+  .option('-e, --env <env>', '构建方式（production||development')
   .option(
     '-l, --link',
     '构建完成后在根目录下的node_modules下创建软链接链接到构建产物方便调试'
@@ -25,7 +25,7 @@ program
 // sand-build start 命令
 program
   .command('start')
-  .description('sand-build start --env <env>')
+  .description('sand-build start --type <type>')
   // build的子命令
   .option('-t, --type <type>', '构建类型（mob||pc||demo）')
   .action(start); // 命中指令后指令build回调