+++ packages/sand-build/src/config/webpack/config/rule/tsLoader.js
@@ -1,0 +1,16 @@
+/**
+ * 获取ts tsx loader
+ */
+function getTsLoader() {
+  return [
+    {
+      test: /\.ts(x?)$/,
+      loader: 'ts-loader',
+      exclude: '/node_modules/',
+    },
+  ];
+}
+
+module.exports = {
+  getTsLoader,
+};