+++ packages/sand-lint/commitlint/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: require.resolve('@commitlint/config-conventional'),
+};