+++ packages/sand-lint/prettier/typescript.js
@@ -1,0 +1,7 @@
+const base = require('./index');
+
+module.exports = {
+  ...base,
+  // ts
+  parser: 'typescript',
+};