+++ staging/sand-demo-staging/.eslintrc.js
@@ -2,7 +2,6 @@
  * eslint配置
  */
 module.exports = {
-  extends: require.resolve('@jianghe/sand-lint/react'),
+  extends: require.resolve('@jianghe/sand-lint/react/prettier'),
   rules: {},
-  settings: {},
 };